use axum's router rather than trait-based lookup and specialize the ticket system trait a bit more.
This PR removes the endpoint trait and replaces it with specified routes for create ticket
and add comment. It also adds some more defined methods to the ticket system trait to make
it clear what should be done in the system with the given payload.

Example logs where we can see the ticket_id in zammad spans:
```
2025-05-10T10:06:39.227912Z  INFO ticket_connector: Listening on http://0.0.0.0:8000/ticket-sync/{create-ticket, add-comment}/<id>
2025-05-10T10:07:35.071304Z  INFO create_ticket{id="1234"}: ticket_connector: creating ticket for ID: 1234
2025-05-10T10:07:35.071603Z  INFO create_ticket{id="1234"}:create_ticket{ticket_id=127}: ticket_connector::models::zammad: Creating Zammad ticket
```